refactor(project): keep detail UI only
