+++ features/project/components/project-detail/index.ts
@@ -1,0 +1,1 @@
+export { ProjectDetailPageContent } from './project-detail-page-content';
