+++ features/project/project.types.ts
@@ -1,7 +1,5 @@
 export type ProjectStatus = 'IN_PROGRESS' | 'COMPLETED' | 'ARCHIVED';
 export type MemberStatus = 'LEADER' | 'MEMBER';
-export type ProjectPosition = string;
-export type ProjectUrlType = 'LINK';
 
 export interface ProjectMemberResponse {
   projectMemberId: number;
@@ -10,7 +8,6 @@ export interface ProjectMemberResponse {
   profileImageUrl: string | null;
   memberStatus: MemberStatus;
   reviewWritten?: boolean;
-  projectPositions?: ProjectPosition[];
 }
 
 export interface ProjectResponse {
@@ -23,10 +20,6 @@ export interface ProjectResponse {
   projectImage: string | null;
   projectStatus: ProjectStatus;
   members: ProjectMemberResponse[];
-  memberCount: number;
-  projectPositions: ProjectPosition[];
-  projectUrl: string;
-  projectUrlType: ProjectUrlType;
 }
 
 /** GET /api/v1/projects — 내 프로젝트 목록 */
@@ -77,10 +70,6 @@ export interface UserProjectResponse {
 export interface ProjectRequest {
   projectName: string;
   host: string;
-  startDate?: string | null;
-  endDate?: string | null;
-  projectUrl?: string;
-  projectUrlType?: ProjectUrlType;
 }
 
 export type ProjectPayload = {
